version 0.12j -------------
Re-added and updated the additional filters under the buy tab so they work on the automated buying option instead of 
just the item clicking , use this with caution , I'd suggest not just filtering by tier etc , but also include partial names
(adept i for example) in the name box.

Added a box to the Craft tab  Box : Enter here the number of the box you want the script to place items in when you use the
'Place items' button on the Main Tab , the script will fill up this box first then attempt to store the rest if it's full
into the other boxes available. (Useful if you use the speciality broker storage boxes).