Examine - Minor fix
Shadow - Removed the code that was turning off eq2bot movement just because shadow was running, this was bad.

Shadow now pauses eq2bot while it is taking control of movement and resumes the script after movement completes.  We'll see if this rapid pause/resume causes any issues or not.  Please provide feedback.