Version 0.14c :  released 31st August 2009
Script went through some MAJOR reworking , reduced the number of variables I used and cleaned up the script a fair bit.

All functions have been tested and still work as they should...if you find any errors , please let me know.

** There IS more script cleanup work to be done sometime in the near future **

Added a filter to the Inventory Tab , enter a phrase or work and press rescan , only the items in your inventory that match that filter will be shown in the list.

Click on an item in your Broker list on the Sell tab and the script will show how many of those items you have on the broker (includes all those items in all boxes)