Version 0.14e :  released 3rd September 2009
Fixed several bugs to do with buying collection items.

1. Duplicate items won't be bought when the name has a comma in it.
2. Fixed the Item Number Counter not decreasing when buying collection items.
3. Went through the script and finally squashed ALL errors to do with item names having commas in them.

The script now keeps track of collection items that have been checked and found to be already collected , it will from then on skip them on the broker , this has more than tripled the speed of scanning for new collection items.