New features
Added BuffBot to script.  This will allow you to monitor tells, group messages,
and raid messages for a trigger phrase to begin buffing.  It can also buff all
those that just rifted.

Also, changed the way of how AutoAttack kicks in when you do not have a melee weapon equiped.  This should fix the random crash to desktop.