Minor Adjustment
doCheckGear implemented to bypass checking gear.  This is one way to guarantee not crashing the script during inventory checks.  I did not add it to the UI so you will need to change the variable in the script before running it.  As it stands, it is currently set to TRUE... thus the way it performs is no different than the previous version.