Improvements
Added a tab called 'Counters' which now gives you full control which ability
you want to counter.  As you identify the target's ability, it will automatically populate the text box so that all you need to do is click the 'Add' button or if you know in advance then you can manually type it in.