Added the following:
AbilityType = this will dump a file listing all abilities by type as well as what information on each piece of equipment

VG-DSC = this is a combat script for the class:  Disciples.  It is currently geared for higher levels but I am pretty sure it will work for lower level characters.