ISXIM - YAHOO INSTANT MESSAGER
Added ISXIM allowing you to use Yahoo Instant Messenger to receive messages as well as send commands to your script.  Fixed a problem with what appears to be the script locking up when you are not looting and the target dies... it now clears the target or fetch the next encounter.