Another improvement:
A) Because of the change in stack quantity to 100, we no longer need to load extra fuel onto the crafting table. 

B) Implemented a routine that will scan the recipe for mandatory ingredients and load them onto the table before doing anything else.  

C) Some work order recipes supplies items such as Leather or Cloth which is not needed on the crafting table.  If your table automatically loads items by default then it will remove your item that matches the supplied item.  This will free up a slot for other items.